Playlist status not being updated on play/pause
Fix issue #4899

The problem was that the playing/paused state was being misreported
to the luahttp interface, turns out this was due to the way the playing
state was being updated in the Qt and hotkey logic.

The p_input state was being updated, but the playlist wasn't being
informed, this patch switches out the p_state calls to use the
equivalent playlist aware functions.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>
(cherry picked from commit 48a98edcc66626a2292f8c9f8d697fa8244736af)

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>