Qt: fix an utf8 encoding issue in the recents menu
The MRL received from a drag & drop was not converted correctly to an
URI (using toURI) before being saved into the configuration. This was
causing multi bytes characters to be serialized in UTF8 (\x????) and be
given as-is to decode_URI which is not able to decode them.
Close #5662

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>
(cherry picked from commit 49571193a5311674483608cd91811ef23f8f0391)

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>