Make getting the time from vlc consistent
VLC has 2 notions of how far into the video it is: time and position.
Time is the absolute time since the beginning of the file. Position is
the the ratio of time and the length of the video. These two *should*
always be consistent by the relationship position = time / video_length.

Currently, time for play/pause command is retrieved directly; time for
seek is calculated as new_position * video_length where new_position is
the new value for position in the position listener.

There is a discrepancy: for the former, time is the *current* time of
where the video is; for the latter, time is where the video will be
after a seek, *but* the seek has not been performed. That is why it used
to be that seeks (on one computer) would match perfectly because both
vlc instances will require a seek, whereas initial sync/play commands
will see a visible delay.

Consistency is number one. Since there is no way to get the "seeked to"
position for play, we make seek use the actual time of the video.

Acutal implementation: add a listener for intf-event which gets called
for all events. Only look at INPUT_EVENT_POSITION. Now when we do a
seek, we set a variable, and the next time INPUT_EVENT_POSITION is
called, we send the seek command. The INPUT_EVENT_POSITION is also the
perfect position to do offline seek.

Note that this actually makes sync visibly worse, but what was working
correctly before was only by a fluke. Once offline sync is implemented,
all should be good...hopefully.