Offline sync!
Relies on having a sync-ed clock so we know where the video should be
compared to the clock of the clock of the player. Then try to catch up
the playback to match that difference. This WILL break playback at
different speeds.

Some parameters are hard coded right now. Works pretty well on my
well-spec laptop with an avi video that has random access and perfect
internet.