Reimplemented the blend filter in C++ and completed the colorspaces supported.
It now supports all combinaisons of
Destinations:
 - YV9, YV12, I410, I411 (8 bits)
 - I420, I422, I444 (8, 9 and 10 bits, native endianness only)
 - YUYV and the various variants
 - NV12/NV21
 - RV15, RV16, RV24, RV32
Sources
 - YUVA
 - RGBA
 - YUVP

The quality of the blending is the same, and the speeds for I420 is the same
(not tested others combinaisons).

Support for blending I420 over another picture has been removed (I don't
see the use case and it simplify the code).

It closes #5477.

(cherry picked from commit fec8f97c7f712483d047e961b31b5cbec921692a)
(cherry picked from commit 4729c93f0cfe187029f0aee1c85253bd54edc2ad)

This commit is backported, because it is the only way to have proper
Hi10p playback on linux using Xv with subtitles. I do not think we want
to deal with the support of people asking why they don't see their subtitles,
and switching to the openGL output with great drivers is...

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>