macosx: fixes aspect ratio / crop functionality in some cases
With this commit the mentioned features result in an actual change, when we are in fullscreen or "macosx-video-autoresize" is false.
Furthermore, this fixes nasty flickering when normal window is changed in size while using this features.
Also this commit makes sure, that no window resize is happening in any case, when "macosx-video-autoresize" is false.

Signed-off-by: Felix Paul Kühne <fkuehne@videolan.org>
(cherry picked from commit fa7456d3cddcfc2ad238ac3102463eef84d134a6)