Fix http live streaming without bandwith set (optional according to spec)
This patch fixes Apple Http Live Streaming of a single stream without bandwith set.
According to the draft RFC, the bandwith parameter is optional. VLC treats this as bandwith zero but
estimates the bandwith based on time it takes to download segments to be able to seek accurately. When
the playlist is updated for a live m3u8, the bandwith value of 0 will cause a new version of the playlist
to be created - effectively stalling the playing stream. This patch makes sure that playlists that do not
specify bandwith reference the same instance.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>
(cherry picked from commit 0b54d2cb5ead026c158556fafc9fea75e8b640f5)

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>