skins2: fix forgotten initialization in copy constructor
This bug was unnoticed because this constructor is only called in a situation
where really copying from origin or leaving it to the compiler to set it via
the default constructor leads to the same result (an empty list). Yet, this
was a bug.

It may solve trac #6599 though,
(a Linux port where there is a real initialization issue)
(cherry picked from commit 076851b74bee39d0c911a344063f5e4731dda28a)

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>