macosx: improve resize behaviour of splitview
Now, the sidebar will not  will not resized while resizing the window.
Also, this commit sets a min width for the the sidebar. But it can
still be hidden, just drag the slider to the left.
(cherry picked from commit 0f4b97b08a1805c6c72f94dbfcbebb61bf956bb8)