macosx: change handling of arrow keys slightly
When the key is assigned to an main menu item: In this case the event
is never passed to the controls, so we want to handle it here for these
exeptional cases.

For all other cases: The event is ignored here and handled by the controls
(playlist, or video view, which also sends it to vlc core).
(cherry picked from commit 0ccbaa821f8b7b928c651df53db5ad50ed246b83)

Signed-off-by: Felix Paul Kühne <fkuehne@videolan.org>