macosx: use a modern system sleep prevention mechanism for the 64bit builds (close #7391)
This is a manual backport of 4a47fcb16. Regrettably, this API was introduced in 10.6 so there is no way to make it work for the legacy 32bit platforms.

The benefit of this approach is that VLC can differenciate between system and screen sleep, so it can allow the system to turn off the screen while playing audio-only media.