macosx: fix display sleep issue on 10.5 (refs #7623)
According to user reports, the modern IOKit approach doesn't work on 10.5 eventhough the docs state that it should.

Note that the screen (and it turn the device) will fall asleep after the defined time in system preferences when playing audio-only media on 32bit devices running 10.5. This is the way it used to be in the past anyway. For x86_64 devices or 32bit devices running 10.6, there is no change.